ACC-129 [ADD] Create Notification Model
