+++ astro/modules/app/base/controllers/base.py
@@ -28,7 +28,6 @@ class BaseController():
 
     def update_all(self):
         json = request.json
-        print(json)
         return self.model.update_all(json=json)
 
     def update(self):
