+++ astro/modules/app/user/models/__init__.py
@@ -1,2 +1,4 @@
+from . import message
 from . import comment
+from . import notification
 from . import user
