+++ astro/modules/app/user/models/user.py
@@ -14,6 +14,7 @@ class User(Base, UserRoute, db.Model, UserMixin):
     password = db.Column(db.String())
     email = db.Column(db.String(120))
     comments = db.relationship("Comment", backref="user", lazy=True)
+    notifications = db.relationship("Notification", backref="user", lazy=True)
 
     def __init__(self) -> None:
         self._controller = UserController(self)
